Separated out authentication functions from the main database file
